Ensure preview functionality is working
